+++ src/egtarget.c
@@ -89,7 +89,7 @@ void updateTargetLock(void) {
         if (idx < 3) {
             lockedRange -= 0x0a00;
         }
-        if (abs(g_ourHead + g_viewHeadingOffset - g_targetBearing) > 0x2000) {
+        if (abs((int16)(g_ourHead + g_viewHeadingOffset - g_targetBearing)) > 0x2000) {
             lockedRange = -32000;
             goto after_lock;
         }
@@ -103,7 +103,7 @@ void updateTargetLock(void) {
     best = -1;
     for (idx = 1; idx < g_planeCount; idx++) {
         computeMapTargetRange(idx);
-        if (abs(g_ourHead + g_viewHeadingOffset - g_targetBearing) < 0x1800 &&
+        if (abs((int16)(g_ourHead + g_viewHeadingOffset - g_targetBearing)) < 0x1800 &&
             idx + 0x80 != g_groundTargetLock && !(g_planeTable.planes[idx].flags & 0x80)) {
             if (g_planeTable.planes[idx].active != 0) {
                 g_targetRange -= 0x280;
@@ -159,7 +159,7 @@ skip_aam:
     if ((g_airTargetLock & 0x80) && g_airTargetLock != -1) {
         idx = g_airTargetLock - 0x80;
         lockedRange = computeTargetBearing(g_simObjects[idx].posX, g_simObjects[idx].posY, 1);
-        if (abs(g_ourHead + g_viewHeadingOffset - g_targetBearing) > 0x2000) {
+        if (abs((int16)(g_ourHead + g_viewHeadingOffset - g_targetBearing)) > 0x2000) {
             lockedRange = 0;
         }
     } else {
@@ -178,7 +178,7 @@ skip_aam:
             !(g_simObjects[idx].flags.b[0] & 0x20) &&
             g_simObjects[idx].speed != 0) {
             computeTargetBearing(g_simObjects[idx].posX, g_simObjects[idx].posY, 1);
-            if (abs(g_ourHead + g_viewHeadingOffset - g_targetBearing) < 0x2000) {
+            if (abs((int16)(g_ourHead + g_viewHeadingOffset - g_targetBearing)) < 0x2000) {
                 range = g_targetRange;
                 best = idx;
             }
@@ -534,7 +534,7 @@ void drawHudWorldOverlay(void) {
                     drawStringActivePage("No Target", 252, 142, 0x0f);
                 }
 
-                if (abs((g_ourHead + g_viewHeadingOffset) - g_targetBearing) > 0x2000) {
+                if (abs((int16)((g_ourHead + g_viewHeadingOffset) - g_targetBearing)) > 0x2000) {
                     g_groundTargetLock = -1;
                 }
             }
@@ -624,7 +624,7 @@ void drawHudWorldOverlay(void) {
     if (g_currentWeaponType == 2 && keyValue == 0) {
         missileSpecD = missiles[missleSpec[missileSpecIndex].weaponIdx].specIndex;
 
-        if (missileSpecD == 30 && abs(g_ourRoll) < 0x2000) {
+        if (missileSpecD == 30 && abs((int16)g_ourRoll) < 0x2000) {
             tmp = computeLoftAngle();
             loftDist = cosMul(tmp, g_altitude) / (sinMul(-tmp, 0x20) + 1);
             pointX = sinMul(g_ourHead, loftDist) + g_viewX_;
